Get or set configuration options that control galah behaviour

The galah package supports large data downloads, and also interfaces with the ALA which requires that users of some services provide a registered email address and reason for downloading data. The ala_config function provides a way to manage these issues as simply as possible.

ala_config(..., preserve = FALSE)

Arguments

...
Options can be defined using the form name = value. Valid arguments are:

reset: ala_config("reset") will reset the options to their default values

caching logical: if TRUE, results will be cached, and any cached results will be re-used). If FALSE, data will be downloaded.

cache_directory string: the directory to use for the cache. By default this is a temporary directory, which means that results will only be cached within an R session and cleared automatically when the user exits R. The user may wish to set this to a non-temporary directory for caching across sessions. The directory must exist on the file system.

download_reason_id numeric or string: the "download reason" required. by some ALA services, either as a numeric ID (currently 0--11) or a string (see find_reasons() for a list of valid ID codes and names). By default this is NA. Some ALA services require a valid download_reason_id code, either specified here or directly to the associated R function.

email string: An email address that has been registered with ALA at this address. A registered email is required for some functions in galah.

send_email logical: should you receive an email for each query to ala_occurrences()? Defaults to FALSE; but can be userful in some instances, for example for tracking DOIs assigned to specific downloads for later citation.

verbose logical: should galah give verbose output to assist debugging? Defaults to FALSE.

preserve
logical: store config for future sessions?

Value

For ala_config(), a list of all options. When ala_config(...) is called with arguments, nothing is returned but the configuration is set.

Examples

if (FALSE) {
 ala_config()
 ala_config(caching = FALSE)
 ala_reasons()
 ala_config(download_reason_id = 0,verbose = TRUE)
}