Count of ALA records

Prior to downloading data it is often valuable to have some estimate of how many records are available, both for deciding if the query is feasible, and for estimating how long it will take to download. Alternatively, for some kinds of reporting, the count of observations may be all that is required, for example for understanding how observations are growing or shrinking in particular locations of for particular taxa. To this end, ala_counts() takes arguments in the same format as ala_occurrences(), and provides either a total count of records matching the criteria, or a data.frame of counts matching the criteria supplied to the group_by argument.

ala_counts(taxa, filters, locations, group_by, limit = 100, type = "record")

Arguments

taxa
data.frame: generated by a call to select_taxa(). This argument also accepts a vector of unique species identifiers.

filters
data.frame: generated by a call to select_filters()

locations
string: generated by a call to select_locations()

group_by
string: field to count by

limit
numeric: maximum number of categories to return, defaulting to 100. If limit is NULL, all results are returned. For some categories this will take a while.

type
string: one of c("record", "species"). Defaults to "record". If "species", the number of species matching the criteria will be returned, if "record", the number of records matching the criteria will be returned.

Value

A single count, if group_by is not specified or,

A data.frame of counts by group_by field, if it is specified

Examples

# With no arguments, return the total number of records in the ALA
ala_counts()

#> [1] 94835639

# Group counts by state and territory
ala_counts(group_by = "stateProvince")

#>                           name    count
#> 1              New South Wales 26256037
#> 2                     Victoria 18715552
#> 3                   Queensland 15658474
#> 4              South Australia  8201923
#> 5            Western Australia  6941269
#> 6           Northern Territory  4807527
#> 7                     Tasmania  4220032
#> 8 Australian Capital Territory  3562577

# Count records matching a filter
ala_counts(filters = select_filters(basisOfRecord = "FossilSpecimen"))

#> [1] 5353

# Count the number of species recorded for each kingdom
if (FALSE) {
ala_counts(group_by = "kingdom", type = "species")
}