Images, sounds and videos

In addition to text data describing individual occurrences and their attributes, ALA stores images, sounds and videos associated with a given record. ala_media allows download of any and all of the media types.

ala_media(
  taxa = NULL,
  filters = NULL,
  locations = NULL,
  columns = select_columns(group = "basic"),
  download_dir,
  refresh_cache = FALSE
)

Arguments

taxa
data.frame: generated by a call to select_taxa(). This argument also accepts a vector of unique species identifiers.

filters
data.frame: generated by a call to select_filters()

locations
string: generated by a call to select_locations()

columns
data.frame: generated by a call to select_columns()

download_dir
string: path to directory to store the downloaded media in

refresh_cache
logical: if set to `TRUE` and `galah_config(caching = TRUE)` then files cached from a previous query will be replaced by the current query

Value

data.frame of metadata of the downloaded media

Details

ala_occurrences() works by first finding all occurrence records matching the filters which contain media, then downloading the metadata for the media and the media files. select_filters() can take both filters relating to occurrences (e.g. basis of records), and filters relating to media (e.g. type of licence). It may be beneficial when requesting a large number of records to show a progress bar by setting verbose = TRUE in galah_config().

See also

Examples

if (FALSE) {
# Download Regent Honeyeater multimedia
media_data <- ala_media(
    taxa = select_taxa("Regent Honeyeater"),
    filters = select_filters(year = 2011),
    download_dir = "media")

# Specify a single media type to download
media_data <- ala_media(
     taxa = select_taxa("Eolophus Roseicapilla"),
     filters = select_filters(multimedia = "Sound"))

# Filter to only records with a particular licence type
media_data <- ala_media(
      taxa = select_taxa("Ornithorhynchus anatinus"),
      filters = select_filters(year = 2020,
      license = "http://creativecommons.org/licenses/by-nc/4.0/")
)
# Check how many records have media files
ala_counts(
     filters = select_filters(multimedia = c("Image","Sound","Video")),
     group_by = "multimedia"
)
}