Occurrence records

The most common form of data stored by ALA are observations of individual life forms, known as 'occurrences'. This function allows the user to search for occurrence records that match their specific criteria, and return them as a data.frame for analysis. Optionally, the user can also request a DOI for a given download to facilitate citation and re-use of specific data resources.

ala_occurrences(
  taxa = NULL,
  filters = NULL,
  locations = NULL,
  columns = select_columns(group = "basic"),
  mint_doi = FALSE,
  doi,
  refresh_cache = FALSE
)

Arguments

taxa
data.frame: generated by a call to select_taxa(). This argument also accepts a vector of unique species identifiers.

filters
data.frame: generated by a call to select_filters()

locations
string: generated by a call to select_locations()

columns
data.frame: generated by a call to select_columns()

mint_doi
logical: by default no DOI will be generated. Set to `TRUE` if you intend to use the data in a publication or similar

doi
string: this argument enables retrieval of occurrence records previously downloaded from the ALA, using the DOI generated by the data.

refresh_cache
logical: if set to `TRUE` and `galah_config(caching = TRUE)` then files cached from a previous query will be replaced by the current query

Value

A data.frame of occurrences, columns as specified by select_columns(). The data.frame object has the following attributes:

a listing of the user-supplied arguments of the data_request (i.e., taxa, filters, locations, columns)

a doi of the data download

the search_url of the query to ALA API

Details

Note that unless care is taken, some queries can be particularly large. While most cases this will simply take a long time to process, if the number of requested records is >50 million the call will not return any data. Users can test whether this threshold will be reached by first calling ala_counts() using the same arguments that they intend to pass to ala_occurrences(). It may also be beneficial when requesting a large number of records to show a progress bar by setting verbose = TRUE in galah_config().

Examples

if (FALSE) {
# Search for occurrences matching a taxon identifier
occ <- ala_occurrences(taxa = select_taxa("Reptilia"))

# Search for occurrences in a year range
occ <- ala_occurrences(filters = select_filters(year = seq(2010, 2020)))

# Search for occurrences in a WKT-specified area
polygon <- "POLYGON((146.24960 -34.05930,146.37045 -34.05930,146.37045 \
-34.152549,146.24960 -34.15254,146.24960 -34.05930))"
occ <- ala_occurrences(locations = select_locations(polygon))
}