Describe what fields are available

Unlike a local tibble, where printing to the console shows what fields are available, when working remotely there is no way to know what fields are present that can be queried. This function name is borrowed from SQL, where it is used to give a read-out of fields that are in the source database.

Usage

describe(x, ...)

# S3 method for class 'data_request'
describe(x, ...)

Arguments

x

An object of class data_request

...

Other arguments, currently ignored

Value

A tibble showing the id, description and data_type for all fields, or if select() is called, then those fields requested by the user. See select() for other examples of valid group values.

See also

Examples

if (FALSE) { # \dontrun{
# By default, this shows all fields in the source system
galah_call() |>
  describe() |>
  collect()

# If `select()` is called, only requested fields are shown
galah_call() |>
  select(group = "basic") |>
  describe() |>
  collect()
} # }