Retrieves field names for use with data retrieval functions

When building a set of filters with select_filters() or specifying required columns with select_columns(), this function can be used to check that the fields map to fields stored in the ALA. This function returns a data.frame of all fields. Field names are in Darwin Core format, except in the case where the field is specific to the ALA database, in which case the ALA field name is returned.

find_fields(class = "all")

Arguments

class
string: class of fields to return e.g. "Assertion". By default all fields are returned.

Value

data.frame of fields with name, data_type, information and Darwin Core class.

Details

The resulting data.frame contains the following columns:

name: Basic description of the field. This is the value that should be used when referring to a field in another function.

data_type: equivalent to the 'class' of a variable in R

info: Additional details about layer

class: Info on the group to which each field belongs

References

Darwin Core terms https://dwc.tdwg.org/terms/

ALA fields https://api.ala.org.au/#ws72

ALA assertion fields https://api.ala.org.au/#ws81

See also

Examples

if (FALSE) {
# Find all fields
 fields <- find_fields()
 # Find assertion fields
 assertions <- find_fields("assertion")
}