Get data filters for a specified data quality profile

Each data quality profile is made up of a series of filters. While some users may wish to simply trust the default filters, it is often useful to check what information they return, particularly if advanced customization is needed. This function gives all of the arguments built into a specific profile.

find_profile_attributes(profile)

Arguments

profile
string: a data quality profile name, short name or id. See find_profiles() for valid filters

Value

A data.frame of profile attributes, consisting of a free text description and the actual filter used.

See also

Examples

profile_info <- find_profile_attributes("CSDM")
profile_info$description # free-text description of each filter in the "CSDM" profile

#> [1] "Include only records where Spatial validity is \"true\""                
#> [2] "Exclude potential duplicate records"                                    
#> [3] "Exclude all records that are an outlier against any environmental layer"
#> [4] "Include only records where Year is 1970 to 2099"