Data quality profiles

The ALA provides a number of pre-built data quality profiles for filtering data according to quality checks. A data quality profile can be specified in the profile argument in select_filters() and used to filter searches in ala_occurrences(), ala_counts() and ala_species().

find_profiles()

Value

A data.frame of available profiles

See also

Examples

# Get available profiles
profile_df <- find_profiles()
# Values given in the 'shortName' column are accepted by select_filter(), i.e.
select_filters(profile = profile_df$shortName[1])

#>                       name include                            value
#>  1:      geospatial_kosher   FALSE                            false
#>  2:             assertions   FALSE                nameNotRecognised
#>  3:             assertions   FALSE            invalidScientificName
#>  4:             assertions   FALSE                     homonymIssue
#>  5:             assertions   FALSE                   unknownKingdom
#>  6:             assertions   FALSE   resourceTaxonomicScopeMismatch
#>  7:             assertions   FALSE                  habitatMismatch
#>  8:             assertions   FALSE             occCultivatedEscapee
#>  9:             assertions   FALSE                     zeroLatitude
#> 10:             assertions   FALSE                    zeroLongitude
#> 11:             assertions   FALSE              invertedCoordinates
#> 12:             assertions   FALSE coordinatesCentreOfStateProvince
#> 13:             assertions   FALSE       coordinatesCentreOfCountry
#> 14:       duplicate_status   FALSE                                D
#> 15: coordinate_uncertainty   FALSE                     [10001 TO *]
#> 16:        user_assertions   FALSE                            50001
#> 17:        user_assertions   FALSE                            50005
#> 18:    outlier_layer_count   FALSE                         [3 TO *]
#> 19:        basis_of_record   FALSE                   FossilSpecimen
#> 20:        basis_of_record   FALSE                 EnvironmentalDNA
#> 21:      occurrence_status   FALSE                           absent
#> 22:                   year   FALSE                      [* TO 1700]
#>                       name include                            value

# is equivalent to:
select_filters(profile = "ALA")

#>                       name include                            value
#>  1:      geospatial_kosher   FALSE                            false
#>  2:             assertions   FALSE                nameNotRecognised
#>  3:             assertions   FALSE            invalidScientificName
#>  4:             assertions   FALSE                     homonymIssue
#>  5:             assertions   FALSE                   unknownKingdom
#>  6:             assertions   FALSE   resourceTaxonomicScopeMismatch
#>  7:             assertions   FALSE                  habitatMismatch
#>  8:             assertions   FALSE             occCultivatedEscapee
#>  9:             assertions   FALSE                     zeroLatitude
#> 10:             assertions   FALSE                    zeroLongitude
#> 11:             assertions   FALSE              invertedCoordinates
#> 12:             assertions   FALSE coordinatesCentreOfStateProvince
#> 13:             assertions   FALSE       coordinatesCentreOfCountry
#> 14:       duplicate_status   FALSE                                D
#> 15: coordinate_uncertainty   FALSE                     [10001 TO *]
#> 16:        user_assertions   FALSE                            50001
#> 17:        user_assertions   FALSE                            50005
#> 18:    outlier_layer_count   FALSE                         [3 TO *]
#> 19:        basis_of_record   FALSE                   FossilSpecimen
#> 20:        basis_of_record   FALSE                 EnvironmentalDNA
#> 21:      occurrence_status   FALSE                           absent
#> 22:                   year   FALSE                      [* TO 1700]
#>                       name include                            value