galah

galah is an R interface to biodiversity data hosted by the Atlas of Living Australia (https://www.ala.org.au/) (ALA). It enables users to locate and download species observations, taxonomic information, or associated media such images or sounds, and to restrict their queries to particular taxa or locations. Users can specify which columns are returned by a query, or restrict their results to observations that meet particular quality-control criteria. All functions return a data.frame as their standard format.

Details

The Atlas of Living Australia is a repository of biodiversity data, focussed primarily on observations of individual life forms. Like the Global Biodiversity Information Facility (GBIF; https://www.gbif.org), the basic unit of data at ALA is an occurrence record, based on the 'Darwin Core' data standard (https://dwc.tdwg.org).

Functions that return data from ALA are named with the prefix ala_, followed by a suffix describing the information that they provide. For example, we anticipate that users will wish to download occurrence data, which can be achieved using the function ala_occurrences(). However, it is also possible to download data on species via ala_species(), or media content (largely images) via ala_media(). Alternatively, users can assess how many records meet their particular criteria using ala_counts().

To get the most value from galah, it is helpful to understand some locally-specific terminology. Each occurrence record contains taxonomic information, and usually some information about the observation itself, such as its location. In addition to this record-specific information, ALA appends contextual information to each record, particularly data from spatial layers reflecting climate gradients or political boundaries (see find_layers() for a full list). Each piece of information associated with a given occurrence is stored in a field (see find_fields()), which corresponds to a column when imported to an R data.frame.

Data fields are important because they provide a means to filter occurrence records; i.e. to return only the information that you need, and no more. Consequently, much of the architecture of galah has been designed to make filtering as simple as possible, by using functions with the select_ prefix. Each select function allows the user to filter in a different way, and again the function suffix contains this information. For example, you can choose which taxonomic groups are included using select_taxa(), or a specific location using select_locations(). By combining different filter functions, it is possible to build complex queries to return only the most valuable information for a given problem.

A notable extention of the filtering approach is to remove records with low 'quality'. ALA performs quality control checks on all records that it stores. These checks are used to generate new fields, that can then be used to filter out records that are unsuitable for particular applications. However, there are many possible data quality checks, and it is not always clear which are most appropriate in a given instance. Therefore, galah supports ALA data quality profiles, which can be passed to select_filters() to quickly remove undesirable records. A full list of data quality profiles is returned by find_profiles().

As the above text makes clear, functions in galah are designed according to a nested architecture. Users that require data should begin by locating the relevant ala_ function; the arguments within that function then call correspondingly-named select_ functions; and finally the specific values that can be interpreted by those select_ functions are given by find_ functions. So, to limit occurrence downloads to a specific taxonomic group, for example, you pass the reult of select_taxa to the taxa argument of ala_occurrences.

Finally, the galah package is named for the bird of the same name (Eolophus roseicapilla), a widely-distributed endemic Australian species.

If you have any questions, comments or suggestions, please email support@ala.org.au.

Functions

Data

ala_occurrences Download occurrence records

ala_species Download species lists

ala_media Download images and sounds

ala_counts Count the number of records returned by a query

Filter

select_taxa Taxon information

select_filters Filter records

select_locations Specify location

select_columns Columns to return in occurrence download

Lookup

find_profiles List data quality profiles

find_profile_attributes List filters included in a data quality profile

search_fields Free-text search for layers and fields

find_fields List occurrence record fields

find_field_values List possible values for a given field

find_layers List spatial layers

find_reasons List valid download reasons

Help

ala_config Package configuration options

ala_citation Citation for a dataset

References

For more information on the ALA API, visit https://api.ala.org.au/