Narrow a query to within a bounding box

Restrict results to within a bounding box (a box constructed from min/max latitude & longitude coordinates). Bounding boxes can be extracted from a supplied sf/sfc object or a shapefile. A bounding box can also be supplied as a bbox object (via sf::st_bbox) or a tibble/data.frame.

Usage

galah_bbox(...)

Arguments

...

an sf object or a shapefile (.shp), or bounding box coordinates supplied as a bbox, a tibble/data.frame

Value

length-1 object of class character and galah_geolocate, containing a multipolygon WKT string representing the bounding box of the area provided.

Details

an sf object or a shapefile polygon will be simplified to its bbox coordinates. A bounding box can be supplied as a bbox object or as a tibble/data.frame. Bounding boxes supplied as a tibble/data.frame must have "xmin", "xmax", "ymin" and "ymax" columns with valid numeric values.

Examples

Search for records using a bounding box of coordinates

b_box <- sf::st_bbox(c(xmin = 143, xmax = 148, ymin = -29, ymax = -28), 
                     crs = st_crs("WGS84"))

galah_call() |>
  galah_identify("reptilia") |>
  galah_bbox(b_box) |>
  atlas_counts()
#> Data returned for bounding box:
#> xmin = 143 xmax = 148 ymin = -29 ymax = -28
#> # A tibble: 1 x 1
#>   count
#>   <int>
#> 1  1782

Search for records using a bounding box in a tibble or data.frame

b_box <- tibble(xmin = 148, ymin = -29, xmax = 143, ymax = -21)

galah_call() |>
  galah_identify("vulpes") |>
  galah_bbox(b_box) |>
  atlas_counts()
#> Data returned for bounding box:
#> xmin = 148 xmax = 143 ymin = -29 ymax = -21
#> # A tibble: 1 x 1
#>   count
#>   <int>
#> 1   225

Search for records within the bounding box of an sf object

galah_config(email = "your-email@email.com")

location <- 
"POLYGON((143.32 -18.78,145.30 -20.52,141.52 -21.50,143.32 -18.78))" |>
 sf::st_as_sfc()
 
galah_call() |>
  galah_identify("vulpes") |>
  galah_bbox(location) |>
  atlas_occurrences()

Search for records within the bounding box of a shapefile

galah_config(email = "your-email@email.com")

location <- sf::st_read(path/to/shapefile.shp)
galah_call() |>
  galah_identify("vulpes") |>
  galah_bbox(location) |>
  atlas_occurrences()

See also