Specify the lowest taxonomic rank required in a downwards search

atlas_taxonomy generates a downwards search of the taxonomic tree. This function can be used to specify the name of a valid taxonomic rank using non-standard evaluation (NSE), for consistency with other galah_ functions.

Usage

galah_down_to(...)

Arguments

...

the name of a single taxonomic rank

Value

A string with the named rank

Examples

An example of using galah_down_to() with atlas_taxonomy(). Return a taxonomic tree of Chordata down to the class level

galah_call() |>
    galah_identify("Chordata") |>
    galah_down_to(class) |>
    atlas_taxonomy()
#>                             levelName
#> 1  Chordata                          
#> 2   ¦--Cephalochordata               
#> 3   ¦   °--Amphioxi                  
#> 4   ¦--Craniata                      
#> 5   ¦   °--Agnatha                   
#> 6   ¦       ¦--Cephalasipidomorphi   
#> 7   ¦       °--Myxini                
#> 8   ¦--Tunicata                      
#> 9   ¦   ¦--Appendicularia            
#> 10  ¦   ¦--Ascidiacea                
#> 11  ¦   °--Thaliacea                 
#> 12  °--Vertebrata                    
#> 13      °--Gnathostomata             
#> 14          ¦--Amphibia              
#> 15          ¦--Aves                  
#> 16          ¦--Mammalia              
#> 17          ¦--Pisces                
#> 18          ¦   ¦--Actinopterygii    
#> 19          ¦   ¦--Chondrichthyes    
#> 20          ¦   ¦--Cephalaspidomorphi
#> 21          ¦   °--Sarcopterygii     
#> 22          °--Reptilia

Another example: return a taxonomic tree of Cacatuidae down to the genus level

galah_call() |>
  galah_identify("Cacatuidae") |>
  galah_down_to(genus) |>
  atlas_taxonomy()
#>                    levelName
#> 1  Cacatuidae               
#> 2   ¦--Cacatuinae           
#> 3   ¦   ¦--Cacatuini        
#> 4   ¦   ¦   ¦--Cacatua      
#> 5   ¦   ¦   ¦--Callocephalon
#> 6   ¦   ¦   ¦--Eolophus     
#> 7   ¦   ¦   °--Lophochroa   
#> 8   ¦   °--Microglossini    
#> 9   ¦       °--Probosciger  
#> 10  ¦--Calyptorhynchinae    
#> 11  ¦   °--Calyptorhynchus  
#> 12  °--Nymphicinae          
#> 13      °--Nymphicus

See also