Narrow a query by passing taxonomic identifiers

When conducting a search or creating a data query, it is common to identify a known taxon or group of taxa to narrow down the records or results returned. identify() is used to identify taxa you want returned in a search or a data query. Users to pass scientific names or taxonomic identifiers with pipes to provide data only for the biological group of interest.

It is good to use search_taxa() and search_identifiers() first to check that the taxa you provide to galah_identify() return the correct results.

Usage

# S3 method for class 'data_request'
identify(x, ...)

# S3 method for class 'metadata_request'
identify(x, ...)

galah_identify(..., search = NULL)

Arguments

x

An object of class metadata_request, created using request_metadata()

...

One or more scientific names.

search

galah_identify() now always does a search to verify search terms; ergo this argument is ignored.

Value

A tibble containing identified taxa.

See also

Examples

if (FALSE) { # \dontrun{
# Use `galah_identify()` to narrow your queries
galah_call() |> 
  identify("Eolophus") |>
  count() |>
  collect()

# If you know a valid taxon identifier, use `filter()` instead.
id <- "https://biodiversity.org.au/afd/taxa/009169a9-a916-40ee-866c-669ae0a21c5c"
galah_call() |> 
  filter(lsid == id)  |>
  count() |>
  collect()
} # }