Search for record information

The living atlases store a huge amount of information, above and beyond the occurrence records that are their main output. In galah, one way that users can investigate this information is by searching for a specific option or category for the type of information they are interested in. Functions prefixed with search_ do this, displaying any matches to a search term within the valid options for the information specified by the suffix.

search_all() is a helper function that can do searches within multiple types of information from search_ sub-functions. See Details (below) for accepted values.

Usage

search_all(type, query)

search_apis(query)

search_assertions(query)

search_atlases(query)

search_collections(query)

search_datasets(query)

search_providers(query)

search_fields(query)

search_licences(query)

search_reasons(query)

search_ranks(query)

search_profiles(query)

search_lists(query)

Arguments

type

A string to specify what type of parameters should be searched.

query

A string specifying a search term. Searches are not case-sensitive.

Value

An object of class tbl_df and data.frame (aka a tibble) containing all data that match the search query.

Details

There are five categories of information, each with their own specific sub-functions to look-up each type of information. The available types of information for search_all() are:

Category	Type	Description	Sub-functions

configuration	atlases	Search for what atlases are available	search_atlases()

apis	Search for what APIs & functions are available for each atlas	search_apis()

reasons	Search for what values are acceptable as 'download reasons' for a specified atlas	search_reasons()

taxonomy	taxa	Search for one or more taxonomic names	search_taxa()

identifiers	Take a universal identifier and return taxonomic information	search_identifiers()

ranks	Search for valid taxonomic ranks (e.g. Kingdom, Class, Order, etc.)	search_ranks()

filters	fields	Search for fields that are stored in an atlas	search_fields()

assertions	Search for results of data quality checks run by each atlas	search_assertions()

licenses	Search for copyright licences applied to media	search_licenses()

group filters	profiles	Search for what data profiles are available	search_profiles()

lists	Search for what species lists are available	search_lists()

data providers	providers	Search for which institutions have provided data	search_providers()

collections	Search for the specific collections within those institutions	search_collections()

datasets	Search for the data groupings within those collections	search_datasets()

Examples

Fields

Search for all fields that use include the word "date"

search_fields("date")
#> # A tibble: 22 x 4
#>   id        description                                              type  link 
#>   <chr>     <chr>                                                    <chr> <chr>
#> 1 month     "Month of observation, specimen collection date. http:/~ fiel~ <NA> 
#> 2 cl23      "Local Government Areas 2012 deprecated LGA Boundaries ~ laye~ http~
#> 3 eventDate "The ISO formatted date of observation, specimen collec~ fiel~ <NA> 
#> 4 cl10903   "Tenure of Australia's forests (2013) v2.0 Tenure of Au~ laye~ http~
#> # ... with 18 more rows

search_all(fields, "date") # equivalent

Search for all fields with the string "basisofrecord"

search_fields("basisofrecord")
#> # A tibble: 2 x 4
#>   id                description                                      type  link 
#>   <chr>             <chr>                                            <chr> <chr>
#> 1 basisOfRecord     What this is a record of e.g. specimen, human o~ fiel~ <NA> 
#> 2 raw_basisOfRecord The basis of record as supplied by the data pub~ fiel~ <NA>

search_all(fields, "basisofrecord") # equivalent

Search for all fields that have information for "marine"

search_fields("marine") |> 
  head() # only show first 5 results
#> # A tibble: 6 x 4
#>   id     description                                                 type  link 
#>   <chr>  <chr>                                                       <chr> <chr>
#> 1 biome  "A field indicating if the record has been located in Terr~ fiel~  <NA>
#> 2 cl2105 "FAO Fishery Statistical Areas FAO Fishing areas (Marine)"  laye~ "htt~
#> 3 el1056 "Endemism (Non-marine) Endemism (Non-marine) based on Cris~ laye~ ""   
#> 4 el957  "Averaged Topographic Relief  This data represents the ave~ laye~ "htt~
#> 5 cl988  "Marine Ecoregions of the World Marine Ecoregions of the W~ laye~ "htt~
#> 6 cl2085 "Global 200 priority ecoregions - Marine WWF's Global 200 ~ laye~ "htt~

search_all(fields, "marine") # equivalent

Search for all Wordclim layers

search_fields("worldclim")
#> # A tibble: 38 x 4
#>   id      description                                                type  link 
#>   <chr>   <chr>                                                      <chr> <chr>
#> 1 el10982 WorldClim 2.1: Temperature - warmest month max Max Temper~ laye~ http~
#> 2 el10981 WorldClim 2.1: Temperature - seasonality Temperature Seas~ laye~ http~
#> 3 el10980 WorldClim 2.1: Temperature - isothermality Isothermality   laye~ http~
#> 4 el10990 WorldClim 2.1: Precipitation - wettest month Precipitatio~ laye~ http~
#> # ... with 34 more rows

search_all(fields, "worldclim") # equivalent

Taxa

Search using a single term (See search_taxa() for more info)

search_taxa("Reptilia")
#> # A tibble: 1 x 9
#>   search_term scientific_name taxon_~1 rank  match~2 kingdom phylum class issues
#>   <chr>       <chr>           <chr>    <chr> <chr>   <chr>   <chr>  <chr> <chr> 
#> 1 Reptilia    REPTILIA        https:/~ class exactM~ Animal~ Chord~ Rept~ noIss~
#> # ... with abbreviated variable names 1: taxon_concept_id, 2: match_type

search_all(taxa, "Reptilia") # equivalent

Note that search_taxa() is not case sensitive

search_taxa("reptilia") # not case sensitive
#> # A tibble: 1 x 9
#>   search_term scientific_name taxon_~1 rank  match~2 kingdom phylum class issues
#>   <chr>       <chr>           <chr>    <chr> <chr>   <chr>   <chr>  <chr> <chr> 
#> 1 reptilia    REPTILIA        https:/~ class exactM~ Animal~ Chord~ Rept~ noIss~
#> # ... with abbreviated variable names 1: taxon_concept_id, 2: match_type

Search multiple taxa. search_taxa() will return one row per taxon

search_taxa(c("reptilia", "mammalia"))
#> # A tibble: 2 x 10
#>   search_term scient~1 taxon~2 rank  match~3 kingdom phylum class issues verna~4
#>   <chr>       <chr>    <chr>   <chr> <chr>   <chr>   <chr>  <chr> <chr>  <chr>  
#> 1 reptilia    REPTILIA https:~ class exactM~ Animal~ Chord~ Rept~ noIss~ <NA>   
#> 2 mammalia    MAMMALIA https:~ class exactM~ Animal~ Chord~ Mamm~ noIss~ Mammals
#> # ... with abbreviated variable names 1: scientific_name, 2: taxon_concept_id,
#> #   3: match_type, 4: vernacular_name

galah_identify() uses search_taxa() to narrow data queries

galah_call() |>
  galah_identify("reptilia") |>
  atlas_counts()
#> # A tibble: 1 x 1
#>     count
#>     <int>
#> 1 1474341

Identifiers

Look up a unique taxon identifier (See search_identifiers() for more info)

search_identifiers(identifier = "https://id.biodiversity.org.au/node/apni/2914510")
#> # A tibble: 1 x 14
#>   scient~1 scien~2 taxon~3 rank  match~4 kingdom phylum class order family genus
#>   <chr>    <chr>   <chr>   <chr> <chr>   <chr>   <chr>  <chr> <chr> <chr>  <chr>
#> 1 Eucalyp~ Maiden  https:~ spec~ taxonI~ Plantae Charo~ Equi~ Myrt~ Myrta~ Euca~
#> # ... with 3 more variables: species <chr>, vernacular_name <chr>,
#> #   issues <chr>, and abbreviated variable names 1: scientific_name,
#> #   2: scientific_name_authorship, 3: taxon_concept_id, 4: match_type

search_all(identifiers, 
           "https://id.biodiversity.org.au/node/apni/2914510") # equivalent

galah_identify() uses identifiers to narrow data queries, too. Ensure that search = FALSE.

galah_call() |>
  galah_identify("https://id.biodiversity.org.au/node/apni/2914510",
                 search = FALSE) |>
  atlas_counts()
#> # A tibble: 1 x 1
#>   count
#>   <int>
#> 1 13466

Lists

Search for species lists that match "endangered"

search_lists("endangered")
#> # A tibble: 7 x 16
#>   dataR~1 listN~2 listT~3 dateC~4 lastU~5 usern~6 fullN~7 itemC~8 region categ~9
#>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <lgl>     <int> <chr>  <chr>  
#> 1 dr19257 Endang~ OTHER   2022-0~ 2022-0~ alanah~ NA           17 <NA>   <NA>   
#> 2 dr4890  Endang~ CONSER~ 2016-0~ 2016-0~ kiashe~ NA          340 <NA>   <NA>   
#> 3 dr17208 My spe~ OTHER   2021-0~ 2021-0~ caterp~ NA            1 <NA>   <NA>   
#> 4 dr17134 Endang~ CONSER~ 2021-0~ 2021-0~ 100588~ NA           20 <NA>   <NA>   
#> 5 dr16896 Rare a~ CONSER~ 2021-0~ 2021-0~ matthe~ NA            1 <NA>   <NA>   
#> 6 dr16203 EPBC A~ SPATIA~ 2020-0~ 2020-0~ robina~ NA           10 <NA>   <NA>   
#> 7 dr19750 Endang~ OTHER   2022-0~ 2022-0~ anneke~ NA           73 <NA>   <NA>   
#> # ... with 6 more variables: generalisation <chr>, authority <chr>,
#> #   sdsType <chr>, isAuthoritative <lgl>, isInvasive <lgl>, isThreatened <lgl>,
#> #   and abbreviated variable names 1: dataResourceUid, 2: listName,
#> #   3: listType, 4: dateCreated, 5: lastUpdated, 6: username, 7: fullName,
#> #   8: itemCount, 9: category

search_all(lists, "endangered") # equivalent

Ranks

Search for a valid taxonomic rank, "subphylum"

search_ranks("subphylum")
#> # A tibble: 1 x 2
#>      id name     
#>   <int> <chr>    
#> 1     7 subphylum

search_all(ranks, "subphylum") # equivalent

Use ranks with galah_down_to() and atlas_taxonomy() to get taxonomic trees

galah_call() %>%
  galah_identify("fungi") %>%
  galah_down_to(subphylum) %>%
  atlas_taxonomy()
#>                        levelName
#> 1  Fungi                        
#> 2   ¦--Ascomycota               
#> 3   ¦   ¦--Pezizomycotina       
#> 4   ¦   ¦--Saccharomycotina     
#> 5   ¦   °--Taphrinomycotina     
#> 6   ¦--Basidiomycota            
#> 7   ¦   ¦--Agaricomycotina      
#> 8   ¦   ¦--Pucciniomycotina     
#> 9   ¦   ¦--Ustilaginomycotina   
#> 10  ¦   °--Wallemiomycotina     
#> 11  ¦--Mucoromycota             
#> 12  ¦   °--Glomeromycotina      
#> 13  °--Zygomycota               
#> 14      ¦--Entomophthoromycotina
#> 15      ¦--Kickxellomycotina    
#> 16      ¦--Mortierellomycotina  
#> 17      ¦--Mucoromycotina       
#> 18      °--Zoopagomycotina

References

Darwin Core terms https://dwc.tdwg.org/terms/

ALA fields https://api.ala.org.au/#ws72

ALA assertions fields https://api.ala.org.au/#ws81

See also