Query layers or fields by free text search

Function to search the ALA database for layers or fields using a free-text search.

search_fields(query, type = "both")

Arguments

query
string: A search string. Not case sensitive.

type
string: What type of parameters should be searched? Should be one of fields, layers or both.

Value

A data.frame with three columns:

id: The identifier for that layer or field. This is the value that should be used when referring to a field in another function.

description: Detailed information on a given field

type: Whether the field is a field or layer

Examples

if (FALSE) {
test <- search_fields("species")
}