Search taxonomic trees

The ALA has its' own internal taxonomy that is derived from authoritative sources. search_taxonomy provides a means to query and visualise that taxonomy, by showing which lower clades are contained within a specified taxon. The inverse query - i.e. listing the higher clades to which a taxon belongs - is provided by select_taxa().

search_taxonomy(query, down_to = NULL)

Arguments

query
string: A vector containing one or more search terms, given as strings. Search terms can be scientific or common names, or taxonomic identifiers. If greater control is required to disambiguate search terms, taxonomic levels can be provided explicitly via a named list for a single name. See vignette(taxonomic_information for using search_taxonomy for multiple clades) Note that searches are not case-sensitive.

down_to
string: A taxonomic rank to search down to. See find_ranks for valid inputs.

Value

a data.frame containing one row per terminal node, and one column for every available intermediate taxonomic level.

Details

The approach used by this function is recursive, meaning that it becomes slow for large queries such as search_taxonomy("Animalia", down_to = "species"). The resulting data.frame can be passed to select_taxa to extract unique identifiers for the terminal nodes (rows), which is necessary to use later functions such as ala_counts or ala_occurrences.

See also

Examples

if (FALSE) {
search_taxonomy("Animalia", down_to = "class")
}