Specify columns for occurrence download

The ALA stores content on hundreds of different fields, and users often thousands or millions of records at a time. To reduce time taken to download data, and limit complexity of the resulting data.frame, it is often sensible to restrict the columns returned by ala_occurrences() to those that are most critical for a given application. This function allows easy selection of individual columns, or commonly-requested groups of columns. The resulting data.frame is then passed to the columns argument in ala_occurrences().

select_columns(..., group)

Arguments

...
zero or more individual column names to include

group
string: (optional) name of one or more column groups to include. Valid options are 'basic', 'event' and 'assertion'

Details

Calling the argument group = 'basic' returns the following columns:

decimalLatitude

decimalLongitude

eventDate

taxon_name

taxonConceptID

recordID

data_resource

Using group = 'event' returns the following columns:

eventRemarks

eventTime

eventID

eventDate

samplingEffort

samplingProtocol

Using group = 'assertion' returns all quality assertion-related columns. The list of assertions is shown by search_fields(type = "assertions").

See also