Select filters to narrow down occurrence queries

'filters' are arguments of the form field logical value that are used to narrow down the number of records returned by a specific query. For example, it is common for users to request records from a particular year (year = 2020), or to return all records except for fossils (basisOfRecord != "FossilSpecimen"). The result of select_filters can be passed to the filters argument in ala_occurrences(), ala_species() or ala_counts().

select_filters(..., profile = NULL)

Arguments

...
filters, in the form field logical value

profile
string: (optional) a data quality profile to apply to the records. See find_profiles for valid profiles. By default no profile is applied.

Value

An object of class data.frame and ala_filters containing filter values.

Details

All statements passed to select_filters() (except the profile argument) take the form of field - logical - value. Permissible examples include:

= or == (e.g. year = 2020)

!=, e.g. year != 2020)

> or >= (e.g. year >= 2020)

< or <= (e.g. year <= 2020)

OR statements (e.g. year == 2018 | year == 2020)

AND statements (e.g. year >= 2000 & year <= 2020)

In some cases R will fail to parse inputs with a single equals sign (=), particularly where statements are separated by /code& or /code|. This problem can be avoided by using a double-equals instead.

See also

Examples

if (FALSE) {
# Create a custom filter for records of interest
filters <- select_filters(
    basisOfRecord = "HumanObservation",
    year >= 2010,
    stateProvince = "New South Wales")

# Add the default ALA data quality profile
filters <- select_filters(
    basisOfRecord = "HumanObservation",
    year >= 2020,
    stateProvince = "New South Wales",
    profile = "ALA")
    
# Use filters to exclude particular values
select_filters(year >= 2010 & year != 2021)

# Separating statements with a comma is equivalent to an 'and' statement, e.g.:
select_filters(year >= 2010 & year < 2020) # is the same as:
select_filters(year >= 2010, year < 2020)

# All statements must include the field name, e.g.
select_filters(year == 2010 | year == 2021) # this works (note double equals)
select_filters(year == 2010 | 2021) # this fails

# solr supports range queries on text as well as numbers, e.g.
select_filters(cl22 >= "Tasmania")
# queries all Australian States & Territories alphabetically after "Tasmania"
}