Taxon information

In the ALA, all records are associated with an identifier that uniquely identifies the taxon to which that record belongs. However, taxonomic names are often ambiguous due to homonymy; i.e. re-use of names (common or scientific) in different clades. Hence, select_taxa provides a means to search for taxonomic names and check the results are 'correct' before proceeded to download data via ala_occurrences(), ala_species() or ala_counts(). The resulting data.frame of taxonomic information can be passed directly to ala_ functions to filter records to the specified taxon or taxa.

select_taxa(
  term,
  term_type = "name",
  return_children = FALSE,
  include_counts = FALSE
)

Arguments

term
string: single string, or optionally a vector, of search term(s). Note that searches are not case-sensitive.

term_type
string: specifies which type of terms are provided in `term`. One of name c("name", "identifier"). Default behaviour is to search by name. identifier refers to the unique identifier for a taxon.

return_children
logical: return child concepts for the provided term(s)?

include_counts
logical: return occurrence counts for all taxa found? FALSE by default

Value

A data.frame of taxonomic information.

See also

Examples

if (FALSE) {
# Search using a single term
select_taxa(term = "Reptilia")
# or equivalently:
select_taxa(term = "reptilia") # not case sensitive

# Search with multiple ranks. This is required if a single term is a homonym.
select_taxa(
  list(kingdom = "Plantae", genus = "Microseris"),
  return_children = TRUE,
  include_counts = TRUE)

# Search using an unique taxon identifier
select_taxa(term = "https://id.biodiversity.org.au/node/apni/2914510",
   term_type = "identifier")

# Search multiple taxa
select_taxa(c("reptilia", "mammalia")) # returns one row per taxon
}