List data quality profiles

The ALA provides a number of pre-built data quality profiles for filtering data according to quality checks. A data quality profile can be specified in the profile argument in galah_filter() and used to filter searches in atlas_occurrences(), atlas_counts() and atlas_species().

Usage

show_all_profiles()

Value

An object of class tbl_df and data.frame (aka a tibble) of available profiles

Examples

Show a list of all available data quality profiles

show_all_profiles()
#> # A tibble: 7 x 4
#>      id name                                                  shortName        description                                 
#>   <int> <chr>                                                 <chr>            <chr>                                       
#> 1     2 ALA General (pre-pipelines)                           ALA-legacy       "The default ALA profile filters out record~
#> 2    35 Species Distribution Modelling (CSDM) (pre-pipelines) CSDM-legacy      "Base filters for the Collaborative Species~
#> 3    44 Data licensed for all uses (pre-pipelines)            re-usable-legacy "Data licensed for re-use, including commer~
#> 4    92 ALA General                                           ALA              "The default ALA profile filters out record~
#> 5   124 Species Distribution Modelling (CSDM)                 CSDM             "Base filters for the Collaborative Species~
#> 6   133 Data licensed for all uses                            re-usable        "Data licensed for re-use, including commer~
#> 7   224 AVH                                                   AVH              "AVH data quality profile"

Values in the shortName column can be used with galah_filter() to narrow your data query results

galah_filter(profile == "ALA")
#> Warning: Invalid field(s) detected.
#> i See a listing of all valid fields with `show_all_fields()`.
#> i Search for the valid name of a desired field with `search_fields()`.
#> x Invalid field(s): profile.
#> # A tibble: 1 x 4
#>   variable logical value query              
#>   <chr>    <chr>   <chr> <chr>              
#> 1 profile  ==      ALA   "(profile:\"ALA\")"

See also