Superseded functions

These functions are still valid, but have been superseded by more recent versions. They were designed to be used to supply arguments within functions from the atlas_ family. Instead you should use piped functions for this same functionality.

Usage

galah_apply_profile(...)

galah_filter(...)

galah_geolocate(..., type = c("polygon", "bbox", "radius"))

galah_polygon(...)

galah_bbox(...)

galah_radius(...)

galah_group_by(...)

galah_identify(...)

galah_select(..., group = NULL)

Arguments

...

Arguments passed to the function in question, usually (but not always) using non-standard evaluation.

type

string: one of "polygon", "bbox" or "radius". Defaults to "polygon". If type = "polygon", a multipolygon will be built via geolocate_polygon(). If type = "bbox", a multipolygon will be built via geolocate_bbox(). The multipolygon is used to narrow a query to the ALA.

group

string: (optional) name of one or more column groups to include. Valid options are "basic", "event" "taxonomy", "media" and "assertions".

Details

Replacements are as follows:

apply_profile() instead of galah_apply_profile()

filter() instead of galah_filter()

geolocate() instead of galah_geolocate()

geolocate_polygon() instead of galah_polygon()

geolocate_bbox() instead of galah_bbox()

geolocate_radius() instead of galah_radius()

group_by() instead of galah_group_by()

identify() instead of galah_identify()

select() instead of galah_select()