Image and sounds

In addition to text data describing individual occurrences and their attributes, ALA stores image and sounds associated with a given record. These can be downloaded to R by first using ala_occurrences() to find records that contain media, and then passing the resulting recordID field to ala_media() to download those media files (see examples).

ala_media(
  identifier,
  identifier_type = "media",
  download_dir,
  media_type = c("image", "sound")
)

Arguments

identifier
string: a single or vector of occurrence or media identifiers. The type is specified by identifier_type

identifier_type
string: one of c("occurrence", "media"). Defaults to "media".

download_dir
string: path to directory to store the downloaded media in

media_type
string: type of media to download, one or both of c("image", "sound"). Defaults to both.

Value

data.frame of media information

Examples

if (FALSE) {
# Search for galah records with images
occ <- ala_occurrences(
    taxa = select_taxa("Eolophus Roseicapilla"),
    filters = select_filters(multimedia = "Image", year = 2020))

# get images associated with these records
images <- ala_media(occ$recordID,
    identifier_type = "occurrence",
    download_dir = tempdir(), media_type = "image")
}