Compute a query

compute() is useful for several purposes. It's original purpose is to send a request for data, which can then be processed by the server and retrieved at a later time (via collect()).

Usage

# S3 method for class 'data_request'
compute(x, ...)

# S3 method for class 'metadata_request'
compute(x, ...)

# S3 method for class 'files_request'
compute(x, ...)

# S3 method for class 'query'
compute(x, ...)

Arguments

x

An object of class data_request, metadata_request or files_request (i.e. constructed using a pipe) or query (i.e. constructed by collapse())

...

Arguments passed on to other methods

Value

An object of class computed_query, which is identical to class query except for occurrence data, where it also contains information on the status of the request.