Query layers, fields or assertions by free text search

This function can be used to find relevant fields and/or layers for use in building a set of filters with select_filters() or specifying required columns with select_columns(). This function returns a data.frame of all fields matching the type specified. Field names are in Darwin Core format, except in the case where the field is specific to the ALA database, in which case the ALA field name is returned.

search_fields(query, type = "all")

Arguments

query
string: A search string. Not case sensitive.

type
string: What type of parameters should be searched? Should be one or more of fields, layers, assertions, or all.

Value

A data.frame with three columns:

id: The identifier for that layer or field. This is the value that should be used when referring to a field in another function.

description: Detailed information on a given field

type: Whether the field is a field or layer

link: For layers, a link to the source data (if available)

Details

Layers are the subset of fields that are spatially appended to each record by the ALA. Layer ids are comprised of a prefix: 'el' for environmental (gridded) layers and 'cl' for contextual (polygon) layers, followed by an id number.

References

Darwin Core terms https://dwc.tdwg.org/terms/

ALA fields https://api.ala.org.au/#ws72

ALA assertions fields https://api.ala.org.au/#ws81

See also

Examples

if (FALSE) {
test <- search_fields("species")

# Find all WorldClim layers
worldclim <- search_fields("worldclim", type = "layers")

# Return a data.frame containing all data on fields and layers
all_fields <- search_fields()
}