Select filters to narrow down occurrence queries

'filters' are arguments of the form field = value that are used to narrow down the number of records returned by a specific query. For example, it is common for users to request records from a particular year (year = 2020), or records that are associated with a physical specimen (basisOfRecord = "PreservedSpecimen"). The result of select_filters can be passed to the filters argument in ala_occurrences(), ala_species() or ala_counts().

select_filters(..., profile = NULL)

Arguments

...
filters, in the form field = value

profile
string: (optional) a data quality profile to apply to the records. See find_profiles for valid profiles. By default no profile is applied.

Value

A data.frame of filter values.

See also

Examples

# Create a custom filter for records of interest
filters <- select_filters(
    basisOfRecord = "HumanObservation",
    year = 2020,
    stateProvince = "New South Wales")

# Add the default ALA data quality profile
filters <- select_filters(
    basisOfRecord = "HumanObservation",
    year = 2020,
    stateProvince = "New South Wales",
    profile = "ALA")