Taxon information

In the ALA, all records are associated with an identifier that uniquely identifies the taxon to which that record belongs. However, taxonomic names are often ambiguous due to homonymy; i.e. re-use of names (common or scientific) in different clades. Hence, select_taxa provides a means to search for taxonomic names and check the results are 'correct' before proceeded to download data via ala_occurrences(), ala_species() or ala_counts(). The resulting data.frame of taxonomic information can be passed directly to ala_ functions to filter records to the specified taxon or taxa.

select_taxa(
  query,
  is_id = FALSE,
  children = FALSE,
  counts = FALSE,
  all_ranks = FALSE
)

Arguments

query
string: A vector containing one or more search terms, given as strings. Search terms can be scientific or common names, or taxonomic identifiers. If greater control is required to disambiguate search terms, taxonomic levels can be provided explicitly via a named list for a single name, or a data.frame for multiple names (see examples). Note that searches are not case-sensitive.

is_id
logical: Is the query a unique identifier? Defaults to FALSE, meaning that queries are assumed to be taxonomic names.

children
logical: Return child concepts for the provided query? DEPRECATED: use search_taxonomy instead.

counts
logical: return occurrence counts for all taxa found? FALSE by default. DEPRECATED: use ala_counts instead.

all_ranks
logical: Include all available intermediate ranks for taxa? e.g. suborder, superfamily. Retrieving this information requires an additional web service call so will slow down the query. DEPRECATED: use search_taxonomy instead.

Value

An object of class data.frame and ala_id containing taxonomic information.

See also

Examples

if (FALSE) {
# Search using a single term
select_taxa("Reptilia")
# or equivalently:
select_taxa("reptilia") # not case sensitive

# Search using an unique taxon identifier
select_taxa(query = "https://id.biodiversity.org.au/node/apni/2914510")

# Search multiple taxa
select_taxa(c("reptilia", "mammalia")) # returns one row per taxon
}