Show valid record information

The living atlases store a huge amount of information, above and beyond the occurrence records that are their main output. In galah, one way that users can investigate this information is by showing all the available options or categories for the type of information they are interested in. Functions prefixed with show_all_ do this, displaying all valid options for the information specified by the suffix.

show_all() is a helper function that can display multiple types of information from show_all_ sub-functions.

Usage

show_all(..., limit = NULL)

show_all_apis(limit = NULL)

show_all_assertions(limit = NULL)

show_all_atlases(limit = NULL)

show_all_collections(limit = NULL)

show_all_datasets(limit = NULL)

show_all_fields(limit = NULL)

show_all_licences(limit = NULL)

show_all_lists(limit = NULL)

show_all_profiles(limit = NULL)

show_all_providers(limit = NULL)

show_all_ranks(limit = NULL)

show_all_reasons(limit = NULL)

Arguments

...

String showing what type of information is to be requested. See Details (below) for accepted values.

limit

Optional number of values to return. Defaults to NULL, i.e. all records

Value

An object of class tbl_df and data.frame (aka a tibble) containing all data of interest.

Details

There are five categories of information, each with their own specific sub-functions to look-up each type of information. The available types of information for show_all_ are:

Category	Type	Description	Sub-functions

configuration	atlases	Show what atlases are available	show_all_atlases()

apis	Show what APIs & functions are available for each atlas	show_all_apis()

reasons	Show what values are acceptable as 'download reasons' for a specified atlas	show_all_reasons()

taxonomy	ranks	Show valid taxonomic ranks (e.g. Kingdom, Class, Order, etc.)	show_all_ranks()

filters	fields	Show fields that are stored in an atlas	show_all_fields()

assertions	Show results of data quality checks run by each atlas	show_all_assertions()

licenses	Show what copyright licenses are applied to media	show_all_licenses()

group filters	profiles	Show what data profiles are available	show_all_profiles()

lists	Show what species lists are available	show_all_lists()

data providers	providers	Show which institutions have provided data	show_all_providers()

collections	Show the specific collections within those institutions	show_all_collections()

datasets	Shows all the data groupings within those collections	show_all_datasets()

References

Darwin Core terms https://dwc.tdwg.org/terms/

See also

Examples

if (FALSE) {
# See all supported atlases
show_all(atlases)

# Show a list of all available data quality profiles
show_all(profiles)

# Show a listing of all accepted reasons for downloading occurrence data
show_all(reasons)

# Show a listing of all taxonomic ranks
show_all(ranks)

# `show_all()` is synonymous with `request_metadata() |> collect()`
request_metadata(type = "fields") |>
  collect()
}